expose more static helpers
