+++ rhino/src/main/java/org/mozilla/javascript/lc/type/TypeInfoFactory.java
@@ -12,6 +12,7 @@ import java.util.Arrays;
 import java.util.Collections;
 import java.util.Date;
 import java.util.EnumSet;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
@@ -134,8 +135,15 @@ public interface TypeInfoFactory extends Serializable {
         return Map.of();
     }
 
-    default List<TypeInfo> consolidateAll(
-            List<TypeInfo> types, Map<VariableTypeInfo, TypeInfo> mapping) {
+    /// helpers
+
+    /**
+     * Transform provided {@code types} by applying {@link TypeInfo#consolidate(Map)} on its elements
+     * <p>
+     * Example: types is {@code [int, E]}, mapping is {@code E -> String}, return value will then be {@code [int, String]}
+     */
+    static List<TypeInfo> consolidateAll(
+        List<TypeInfo> types, Map<VariableTypeInfo, TypeInfo> mapping) {
         if (mapping.isEmpty()) { // implicit null check
             return types;
         }
@@ -159,7 +167,25 @@ public interface TypeInfoFactory extends Serializable {
         return consolidatedTypes;
     }
 
-    /// helpers
+    /**
+     * Transform a mapping by applying {@link TypeInfo#consolidate(Map)} on its values.
+     * <p>
+     * Example: mapping is {@code K -> V}, transformer is {@code V -> String}, return value will then be {@code K -> String}
+     */
+    static Map<VariableTypeInfo, TypeInfo> transformMapping(
+        Map<VariableTypeInfo, TypeInfo> mapping, Map<VariableTypeInfo, TypeInfo> transformer) {
+        if (mapping.isEmpty()) {
+            return Map.of();
+        } else if (mapping.size() == 1) {
+            var entry = mapping.entrySet().iterator().next();
+            return Map.of(entry.getKey(), entry.getValue().consolidate(transformer));
+        }
+        var transformed = new HashMap<>(mapping);
+        for (var entry : transformed.entrySet()) {
+            entry.setValue(entry.getValue().consolidate(transformer));
+        }
+        return transformed;
+    }
 
     /**
      * create a {@link TypeInfo} from {@link Type}.
