+++ rhino/src/main/java/org/mozilla/javascript/lc/type/impl/factory/FactoryBase.java
@@ -59,21 +59,6 @@ public interface FactoryBase extends TypeInfoFactory {
         return new ParameterizedTypeInfoImpl(base, params);
     }
 
-    private static Map<VariableTypeInfo, TypeInfo> transformMapping(
-            Map<VariableTypeInfo, TypeInfo> mapping, Map<VariableTypeInfo, TypeInfo> transformer) {
-        if (mapping.isEmpty()) {
-            return Map.of();
-        } else if (mapping.size() == 1) {
-            var entry = mapping.entrySet().iterator().next();
-            return Map.of(entry.getKey(), entry.getValue().consolidate(transformer));
-        }
-        var transformed = new HashMap<>(mapping);
-        for (var entry : transformed.entrySet()) {
-            entry.setValue(entry.getValue().consolidate(transformer));
-        }
-        return transformed;
-    }
-
     /** Used by {@link #getConsolidationMapping(java.lang.Class)} */
     default Map<VariableTypeInfo, TypeInfo> computeConsolidationMapping(Class<?> type) {
         var mapping = new HashMap<VariableTypeInfo, TypeInfo>();
@@ -105,9 +90,9 @@ public interface FactoryBase extends TypeInfoFactory {
         // then merge them together
         // Example: Ta -> Tb (from `superMapping`) will be transformed by Tb -> Te (from `mapping`),
         // forming Ta -> Te
-        var merged = new HashMap<>(transformMapping(superMapping, mapping));
+        var merged = new HashMap<>(TypeInfoFactory.transformMapping(superMapping, mapping));
         for (var interfaceMapping : interfaceMappings) {
-            merged.putAll(transformMapping(interfaceMapping, mapping));
+            merged.putAll(TypeInfoFactory.transformMapping(interfaceMapping, mapping));
         }
         merged.putAll(mapping);
